persist: move LeasedBatch into persist-client::fetch
