+++ src/persist-client/src/read.rs
@@ -27,12 +27,11 @@ use uuid::Uuid;
 use mz_persist::location::{Blob, SeqNo};
 use mz_persist_types::{Codec, Codec64};
 
-use crate::fetch::{fetch_batch, BatchFetcher};
-pub use crate::internal::encoding::SerdeLeasedBatch;
+use crate::fetch::{fetch_batch, BatchFetcher, LeaseLifeCycle, LeasedBatch};
 use crate::internal::machine::Machine;
 use crate::internal::metrics::Metrics;
-use crate::internal::state::{HollowBatch, Since};
-use crate::{GarbageCollector, PersistConfig, ShardId};
+use crate::internal::state::Since;
+use crate::{GarbageCollector, PersistConfig};
 
 /// An opaque identifier for a reader of a persist durable TVC (aka shard).
 #[derive(Clone, PartialEq, Eq, Hash)]
@@ -53,181 +52,6 @@ impl std::fmt::Debug for ReaderId {
 impl ReaderId {
     pub(crate) fn new() -> Self {
         ReaderId(*Uuid::new_v4().as_bytes())
-    }
-}
-
-/// Propagates metadata from readers alongside a `HollowBatch` to apply the
-/// desired semantics.
-#[derive(Debug, Clone)]
-pub(crate) enum LeasedBatchMetadata<T> {
-    /// Apply snapshot-style semantics to the fetched batch.
-    Snapshot {
-        /// Return all values with time leq `as_of`.
-        as_of: Antichain<T>,
-    },
-    /// Apply listen-style semantics to the fetched batch.
-    Listen {
-        /// Return all values with time in advance of `as_of`.
-        as_of: Antichain<T>,
-        /// Return all values with time leq `until`.
-        until: Antichain<T>,
-    },
-}
-
-#[derive(Debug)]
-pub(crate) enum LeaseLifeCycle {
-    Issued { seqno: SeqNo, droppable: bool },
-    Consumed { seqno: SeqNo, droppable: bool },
-    Completed,
-}
-
-impl LeaseLifeCycle {
-    fn seqno(&self) -> Option<SeqNo> {
-        use LeaseLifeCycle::*;
-        match self {
-            Issued { seqno, .. } | Consumed { seqno, .. } => Some(*seqno),
-            Completed => None,
-        }
-    }
-}
-
-/// A token representing one read batch.
-///
-/// This may be exchanged (including over the network). It is tradeable via
-/// [`crate::fetch::fetch_batch`] for the resulting data stored in the batch.
-///
-/// # Panics
-/// `LeasedBatch` panics when dropped unless a very strict set of invariants are
-/// held:
-///
-/// `LeasedBatch` may only be dropped if it:
-/// - Is being converted to `SerdeLeasedBatch`.
-/// - Has been returned to its issuer via `Self::return_lease`.
-///
-/// In any other circumstance, dropping `LeasedBatch` panics.
-#[derive(Debug)]
-pub struct LeasedBatch<T>
-where
-    T: Timestamp + Codec64,
-{
-    pub(crate) shard_id: ShardId,
-    pub(crate) reader_id: ReaderId,
-    pub(crate) metadata: LeasedBatchMetadata<T>,
-    pub(crate) batch: HollowBatch<T>,
-    /// The `SeqNo` from which this batch originated; we track this value as
-    /// long as necessary to ensure the `SeqNo` isn't garbage collected while a
-    /// read still depends on it.
-    pub(crate) leased_seqno: LeaseLifeCycle,
-}
-
-impl<T> LeasedBatch<T>
-where
-    T: Timestamp + Codec64,
-{
-    pub(crate) fn new(
-        shard_id: ShardId,
-        reader_id: ReaderId,
-        metadata: LeasedBatchMetadata<T>,
-        batch: HollowBatch<T>,
-        leased_seqno: LeaseLifeCycle,
-    ) -> LeasedBatch<T> {
-        LeasedBatch {
-            shard_id,
-            reader_id,
-            metadata,
-            batch,
-            leased_seqno,
-        }
-    }
-
-    /// Takes `self` into a [`SerdeLeasedBatch`], allowing users to drop `self`.
-    ///
-    /// !!!WARNING!!!
-    ///
-    /// If `self` has a `leased_seqno`, failing to take the returned
-    /// `SerdeLeasedBatch` back into a `LeasedBatch` will leak `SeqNo`s and
-    /// prevent persist compaction.
-    pub fn get_droppable_batch(mut self) -> SerdeLeasedBatch {
-        // Making `LeasedBatch`es with outstanding leases droppable should only
-        // occur here; this should not be a generally accessible function
-        match &mut self.leased_seqno {
-            // Leases can only be marked completed on their issuer, and they
-            // should not be exchanged afterward.
-            LeaseLifeCycle::Completed => {}
-            LeaseLifeCycle::Issued { droppable, .. }
-            | LeaseLifeCycle::Consumed { droppable, .. } => *droppable = true,
-        }
-
-        SerdeLeasedBatch::from(self)
-    }
-
-    /// Signals whether or not `self` should downgrade the `Capability` its
-    /// presented alongside.
-    pub fn generate_progress(&self) -> Option<Antichain<T>> {
-        match self.metadata {
-            LeasedBatchMetadata::Listen { .. } => Some(self.batch.desc.upper().clone()),
-            LeasedBatchMetadata::Snapshot { .. } => None,
-        }
-    }
-
-    /// Identifies that the batch has been consumed. After calling this
-    /// function, the batch must either be returned to its issuer, or promise a
-    /// return.
-    ///
-    /// # Panics
-    /// - If lease has already been consumed
-    fn consume_lease(&mut self) {
-        match &self.leased_seqno {
-            LeaseLifeCycle::Issued {
-                seqno,
-                droppable: false,
-            } => {
-                self.leased_seqno = LeaseLifeCycle::Consumed {
-                    seqno: *seqno,
-                    droppable: false,
-                }
-            }
-            _ => panic!("each lease may be consumed at most once"),
-        }
-    }
-
-    /// Because sources get dropped without notice, we need to permit another
-    /// operator to safely expire leases.
-    ///
-    /// The batch's `reader_id` is intentionally inaccessible, and should be
-    /// obtained from the issuing [`ReadHandle`], or one of its derived
-    /// structures, e.g. [`Subscribe`].
-    ///
-    /// # Panics
-    /// - If `reader_id` is different than the [`ReaderId`] from the batch
-    ///   issuer.
-    fn return_lease(&mut self, reader_id: &ReaderId) -> Option<SeqNo> {
-        assert!(
-            &self.reader_id == reader_id,
-            "only issuing reader can authorize lease expiration"
-        );
-        let seqno = self.leased_seqno.seqno();
-        self.leased_seqno = LeaseLifeCycle::Completed;
-        seqno
-    }
-}
-
-impl<T> Drop for LeasedBatch<T>
-where
-    T: Timestamp + Codec64,
-{
-    /// For details, see [`LeasedBatch`].
-    fn drop(&mut self) {
-        assert!(
-            match self.leased_seqno {
-                LeaseLifeCycle::Completed => true,
-                LeaseLifeCycle::Issued { droppable, .. }
-                | LeaseLifeCycle::Consumed { droppable, .. } => droppable,
-            },
-            "LeasedBatch must be explicitly permitted to drop or its \
-            lease must have been returned. This batch was not: {:?}",
-            self
-        );
     }
 }
 
@@ -424,7 +248,7 @@ where
         let r = LeasedBatch {
             shard_id: self.handle.machine.shard_id(),
             reader_id: self.handle.reader_id.clone(),
-            metadata: LeasedBatchMetadata::Listen {
+            metadata: crate::fetch::LeasedBatchMetadata::Listen {
                 as_of: self.as_of.clone(),
                 until: self.frontier.clone(),
             },
@@ -649,7 +473,7 @@ where
             .map(|batch| LeasedBatch {
                 shard_id: self.machine.shard_id(),
                 reader_id: self.reader_id.clone(),
-                metadata: LeasedBatchMetadata::Snapshot {
+                metadata: crate::fetch::LeasedBatchMetadata::Snapshot {
                     as_of: as_of.clone(),
                 },
                 batch,
